enhanced_agents/requirements.txt
# LangChain Enhanced Agents Dependencies
# 核心LangChain包
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# LLM 提供商
langchain-openai>=0.0.8
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6

# Web界面
gradio>=4.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
faiss-cpu>=1.7.4

# HTTP请求
requests>=2.28.0
httpx>=0.25.0

# 异步支持
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# 日志和配置
python-dotenv>=1.0.0
pyyaml>=6.0

# 数据科学工具（可选）
matplotlib>=3.6.0
plotly>=5.0.0
scikit-learn>=1.3.0

# 开发工具
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# 向量数据库（可选）
chromadb>=0.4.0
weaviate-client>=3.24.0

# 文档处理（可选）
pypdf2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# 工具扩展（可选）
google-search-results>=2.4.0
wikipedia>=1.4.0